+++ R/initial_search.R
@@ -47,7 +47,8 @@ Jaatha.initialSearch <- function(jaatha, sim=200, blocks.per.par=2, rerun=FALSE)
   firstBlocks <- createInitialBlocks(jaatha@par.ranges, blocks.per.par)
 
   for (i in seq(along=firstBlocks)){
-    .print("*** Block", i, "of", length(firstBlocks), ":", printBorder(firstBlocks[[i]], jaatha))
+    .print("*** Block", i, "of", length(firstBlocks), ":", 
+           printBorder(firstBlocks[[i]], jaatha))
 
     sim.data <- list()     
     
@@ -59,7 +60,6 @@ Jaatha.initialSearch <- function(jaatha, sim=200, blocks.per.par=2, rerun=FALSE)
         suppressWarnings( glms.fitted <- fitGlm(sim.data, jaatha) )
         break
       }, error = function(e) {
-        .log2("Error fitting GLM:", e$message)
         if (j < 5) .print("Failed to fit the GLM. Retrying with more simulations...")
         else stop('Failed to fit the GLM. Try disabeling smoothing or using more simulations')
       })